Add simple movement.

+++ inc/clock.h
@@ -15,7 +15,7 @@ typedef struct
 
 Clock *clock_init();
 void   clock_tick(Clock *clock);
-size_t clock_dt(Clock *clock);
+float  clock_dt(Clock *clock);
 size_t clock_fps(Clock *clock);
 
 #endif /* CLOCK_H_ */
